+++ crates/node-api/src/evm/mod.rs
@@ -6,7 +6,3 @@ pub trait EvmConfig: ConfigureEvmEnv + Clone + Send + Sync + 'static {}
 /// Traits for working with a configurable EVM.
 mod traits;
 pub use traits::ConfigureEvmEnv;
-
-/// Bundle state with receipts.
-mod bundle_state_with_receipts;
-pub use bundle_state_with_receipts::*;
