+++ crates/storage/provider/src/test_utils/executor.rs
@@ -1,6 +1,6 @@
 use crate::{
-    bundle_state::BundleStateWithReceipts, BlockExecutor, BlockExecutorStats, ExecutorFactory,
-    PrunableBlockExecutor, StateProvider,
+    bundle_state::BundleStateWithReceipts, BlockExecutor, ExecutorFactory, PrunableBlockExecutor,
+    StateProvider,
 };
 use parking_lot::Mutex;
 use reth_interfaces::executor::BlockExecutionError;
@@ -11,6 +11,8 @@ use std::sync::Arc;
 pub struct TestExecutor(pub Option<BundleStateWithReceipts>);
 
 impl BlockExecutor for TestExecutor {
+    type Error = BlockExecutionError;
+
     fn execute(
         &mut self,
         _block: &BlockWithSenders,
@@ -45,10 +47,6 @@ impl BlockExecutor for TestExecutor {
         self.0.clone().unwrap_or_default()
     }
 
-    fn stats(&self) -> BlockExecutorStats {
-        BlockExecutorStats::default()
-    }
-
     fn size_hint(&self) -> Option<usize> {
         None
     }
@@ -77,7 +75,7 @@ impl ExecutorFactory for TestExecutorFactory {
     fn with_state<'a, SP: StateProvider + 'a>(
         &'a self,
         _sp: SP,
-    ) -> Box<dyn PrunableBlockExecutor + 'a> {
+    ) -> Box<dyn PrunableBlockExecutor<Error = <TestExecutor as BlockExecutor>::Error> + 'a> {
         let exec_res = self.exec_results.lock().pop();
         Box::new(TestExecutor(exec_res))
     }
