+++ crates/storage/provider/src/traits/executor.rs
@@ -6,9 +6,6 @@ use reth_primitives::{BlockNumber, BlockWithSenders, PruneModes, Receipt, U256};
 use std::time::Duration;
 use tracing::debug;
 
-/// re-export node-api BlockExecutor for now
-pub use reth_node_api::evm::BlockExecutor;
-
 /// Executor factory that would create the EVM with particular state provider.
 ///
 /// It can be used to mock executor.
@@ -17,7 +14,53 @@ pub trait ExecutorFactory: Send + Sync + 'static {
     fn with_state<'a, SP: StateProvider + 'a>(
         &'a self,
         _sp: SP,
-    ) -> Box<dyn PrunableBlockExecutor + 'a>;
+    ) -> Box<dyn PrunableBlockExecutor<Error = BlockExecutionError> + 'a>;
+}
+
+/// An executor capable of executing a block.
+pub trait BlockExecutor {
+    /// The error type returned by the executor.
+    type Error;
+
+    /// Execute a block.
+    fn execute(
+        &mut self,
+        block: &BlockWithSenders,
+        total_difficulty: U256,
+    ) -> Result<(), Self::Error>;
+
+    /// Executes the block and checks receipts.
+    ///
+    /// See [execute](BlockExecutor::execute) for more details.
+    fn execute_and_verify_receipt(
+        &mut self,
+        block: &BlockWithSenders,
+        total_difficulty: U256,
+    ) -> Result<(), Self::Error>;
+
+    /// Runs the provided transactions and commits their state to the run-time database.
+    ///
+    /// The returned [BundleStateWithReceipts] can be used to persist the changes to disk, and
+    /// contains the changes made by each transaction.
+    ///
+    /// The changes in [BundleStateWithReceipts] have a transition ID associated with them: there is
+    /// one transition ID for each transaction (with the first executed tx having transition ID
+    /// 0, and so on).
+    ///
+    /// The second returned value represents the total gas used by this block of transactions.
+    ///
+    /// See [execute](BlockExecutor::execute) for more details.
+    fn execute_transactions(
+        &mut self,
+        block: &BlockWithSenders,
+        total_difficulty: U256,
+    ) -> Result<(Vec<Receipt>, u64), Self::Error>;
+
+    /// Return bundle state. This is output of executed blocks.
+    fn take_output_state(&mut self) -> BundleStateWithReceipts;
+
+    /// Returns the size hint of current in-memory changes.
+    fn size_hint(&self) -> Option<usize>;
 }
 
 /// A [BlockExecutor] capable of in-memory pruning of the data that will be written to the database.
